Remove magic numbers from Mario
Replace unexplained 50, 24, 62 with semantically meaningful values. This change was inspired by the debugger - changing either of the new parameters actually makes sense in the drawing.

Also change heading to `INPUT` as described in "accidentally well-architected code".

In theory, `marioH` should be half the height of the gif, but his feet aren't on the bottom of the image.