Syntax.elm: Use different example for defining own infix operators
Hitherto `<|` is used as an user defined operator. But this is a builtin operator that can't be overwritten.
So I propose to use `?` as an infix variant of Maybe.withDefault.

Code for the function is taken from @evancz at https://groups.google.com/d/msg/elm-discuss/kgiM9j3Xrbw/EA9KSSqNkBYJ