Different variable names from type variables
I am writing a small ELM beginners guide. I am at the part of explaining pattern matching. While trying to explain this example, I feel that some people might confuse `a` and `b` from the type signature (`Either a b`) with this `a` and `b` from the pattern matches. I tried to come up with a more descriptive name, just to avoid an unnecessary clash.