Add link to companion blog post
I tried to add an elm component in to a React project, and found it surprisingly difficult (I almost gave up), due to a variety of small niggles. Generally, when I find something difficult and there isn't good help on the internet, I tend to write a blog post about it.

This commit adds a link to a blog post that has detailed step by step instructions, alongside a GitHub repo with corresponding commits. I hope that following this blog post will turn what was for me a long job of exploration in to a quick and easy job.