Fix drag and drop example
This makes the text in the draggable box non-selectable,
which prevents issue with dragged box not being released
in the rare cases when you inadvertently select the text in the box
and then start dragging by clicking within the selected text.
In such case dragging is not stopped by releasing mouse button and
the example looks broken.