Make time example a bit more intuitive.
I think this approach might be better. It assumes some familiarity with trigonometry but so does using `sin` and `cos` further down. It also avoids the use of `turns` which was not introduced and it does not call `inMinutes`. The SVG shows the seconds hand so it is a bit confusing to call `inMinutes` in the source code.